Only run pure unit tests + always prebuild but only upload when version changed
