+++ jenkins/Jenkinsfile.linux
@@ -12,13 +12,10 @@ pipeline {
         sh 'npm run reset'
         sh 'npm run submodule'
         sh 'npm install --verbose'
+        sh 'ldd build/Release/deltachat.node'
       }
     }
     stage('Tests') {
-      environment {
-        DC_ADDR = credentials('dc-addr')
-        DC_MAIL_PW = credentials('dc-mail-pw')
-      }
       steps {
         sh 'npm test'
       }
@@ -29,6 +26,8 @@ pipeline {
       }
       steps {
         sh 'npm run prebuild'
+        sh 'ldd prebuilds/linux-x64/node-napi.node'
+        sh 'ldd prebuilds/linux-x64/electron-napi.node'
       }
     }
   }
