+++ package.json
@@ -5,10 +5,10 @@
   "scripts": {
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "coverage-html-report": "rm -rf coverage/ && nyc report --reporter=html && opn coverage/index.html",
-    "dependency-check": "dependency-check package.json --missing --unused -i dependency-check -i version-changed -i nyc -i opn-cli -i standard -i napi-macros -i coveralls --entry index.js --entry test/index.js --entry scripts/*.js",
+    "dependency-check": "dependency-check package.json --missing --unused -i dependency-check -i nyc -i opn-cli -i standard -i napi-macros -i coveralls --entry index.js --entry test/index.js --entry scripts/*.js",
     "generate-constants": "./scripts/generate-constants.js",
     "install": "node-gyp-build scripts/rebuild-core.js",
-    "prebuild": "version-unchanged || node scripts/prebuildify.js",
+    "prebuild": "node scripts/prebuildify.js",
     "submodule": "git submodule update --recursive --init",
     "test": "standard && npm run dependency-check && nyc node test/index.js",
     "reset": "rm -rf node_modules/ build/ prebuilds/ deltachat-core/"
